[task/140371] Refactoring API layer
