+++ src/wsm/wsm/utils/AppConstant.swift
@@ -9,6 +9,9 @@
 import UIKit
 
 class AppConstant {
+    public static let debugAPIBaseUrl = "http://edev.framgia.vn"
+    public static let productAPIBaseUrl = "http://wsm.framgia.vn"
+
     public static let DateFormatEEEDisplay = "EEE, d MMM yyyy"
     public static let TimeFormat24HDisplay = "HH:mm"
 }
